feat: Auth por JWT utilizando tabela Usuario(login, password) como dados de login.

+++ src/main/java/backend/dev_mobile/my_economy/model/entity/Despesa.java
@@ -1,7 +1,6 @@
 package backend.dev_mobile.my_economy.model.entity;
 
 import java.time.LocalDate;
-import java.time.YearMonth;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -15,7 +14,7 @@ public class Despesa {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String descricao;
     private Double quantidade;
     private LocalDate referenciaMes;
